+++ .gitignore
@@ -1,4 +1,5 @@
 .next
+.idea
 node_modules
 .env
 .DS_Store
