+++ components/containers/Container.js
@@ -1,4 +1,4 @@
-import containerStyles from '../styles/Container.module.scss';
+import containerStyles from '../../styles/Container.module.scss';
 
 export default function Container({ className, children, styles }) {
   return (
