+++ pages/_app.js
@@ -1,4 +1,4 @@
-import Layout from '../components/Layout';
+import Layout from '../components/layout/Layout';
 import '../styles/globals.scss';
 
 function MyApp({ Component, pageProps }) {
